refactor(frontend): data selection state
Remove `state/data-selection`. Move data selection Recoil state for individual data layers to corresponding folders in `data-layers` directory. Move hazard damage mapping state to `state/damage-mapping`.
