+++ frontend/src/data-layers/buildings/data-selection.ts
