+++ frontend/src/data-layers/buildings/state.ts
@@ -3,8 +3,9 @@ import { selector } from 'recoil';
 import { ViewLayer } from 'lib/data-map/view-layers';
 import { truthyKeys } from 'lib/helpers';
 import { buildingsViewLayer } from 'data-layers/buildings/buildings-view-layer';
-import { buildingSelectionState } from 'state/data-selection/buildings';
 import { sectionVisibilityState } from 'state/sections';
+
+import { buildingSelectionState } from './data-selection';
 
 export const buildingsLayerState = selector<ViewLayer[]>({
   key: 'buildingLayersState',
