+++ frontend/src/data-layers/hazards/data-selection.ts
