+++ frontend/src/data-layers/networks/networks-style.ts
