+++ frontend/src/sidebar/sections/buildings/BuildingsControl.tsx
@@ -1,8 +1,9 @@
+import { useRecoilState } from 'recoil';
+
+import { buildingSelectionState } from 'data-layers/buildings/data-selection';
 import { NETWORKS_METADATA } from 'data-layers/networks/metadata';
 import { ParamChecklist } from 'lib/controls/params/ParamChecklist';
-import { useRecoilState } from 'recoil';
 import { LayerLabel } from 'sidebar/ui/LayerLabel';
-import { buildingSelectionState } from 'state/data-selection/buildings';
 
 export const BuildingsControl = () => {
   const [checkboxState, setCheckboxState] = useRecoilState(buildingSelectionState);
