+++ frontend/src/state/damage-mapping/damage-style-params.ts
