+++ frontend/src/state/layers/ui/labels.ts
@@ -3,7 +3,7 @@ import { selector } from 'recoil';
 import { ViewLayer, viewOnlyLayer } from 'lib/data-map/view-layers';
 import { regionLabelsDeckLayer } from 'data-layers/regions/region-labels-deck-layer';
 import { sectionVisibilityState } from 'state/sections';
-import { regionLevelState } from 'state/data-selection/regions';
+import { regionLevelState } from 'data-layers/regions/data-selection';
 import { backgroundState, showLabelsState } from 'map/layers/layers-state';
 
 export const labelsLayerState = selector<ViewLayer[]>({
